move dataset to ego file
