+++ l5kit/l5kit/dataset/__init__.py
@@ -1,6 +1,5 @@
 from .agent import AgentDataset
-from .ego import BaseEgoDataset, EgoDataset
-from .ego_vectorized import EgoDatasetVectorized
+from .ego import BaseEgoDataset, EgoDataset, EgoDatasetVectorized
 from .select_agents import select_agents
