+++ l5kit/l5kit/dataset/ego.py
@@ -5,11 +5,13 @@ from typing import Callable, Optional
 import numpy as np
 from torch.utils.data import Dataset
 
-from ..data import ChunkedDataset, get_frames_slice_from_scenes
-from ..kinematic import Perturbation
-from ..rasterization import Rasterizer, RenderContext
-from ..sampling import generate_agent_sample
-from .utils import convert_str_to_fixed_length_tensor
+from l5kit.data import ChunkedDataset, get_frames_slice_from_scenes
+from l5kit.dataset.utils import convert_str_to_fixed_length_tensor
+from l5kit.kinematic import Perturbation
+from l5kit.rasterization import Rasterizer, RenderContext
+from l5kit.sampling.agent_sampling import generate_agent_sample
+from l5kit.sampling.agent_sampling_vectorized import generate_agent_sample_vectorized
+from l5kit.vectorization.vectorizer import Vectorizer
 
 
 class BaseEgoDataset(Dataset):
@@ -205,3 +207,42 @@ class EgoDataset(BaseEgoDataset):
         """
         dataset = self.dataset.get_scene_dataset(scene_index)
         return EgoDataset(self.cfg, dataset, self.rasterizer, self.perturbation)
+
+
+class EgoDatasetVectorized(BaseEgoDataset):
+    def __init__(
+        self,
+        cfg: dict,
+        zarr_dataset: ChunkedDataset,
+        vectorizer: Vectorizer,
+        perturbation: Optional[Perturbation] = None,
+    ):
+        """
+        Get a PyTorch dataset object that can be used to train DNNs with vectorized input
+
+        Args:
+            cfg (dict): configuration file
+            zarr_dataset (ChunkedDataset): the raw zarr dataset
+            vectorizer (Vectorizer): a object that supports vectorization around an AV
+            perturbation (Optional[Perturbation]): an object that takes care of applying trajectory perturbations.
+        None if not desired
+        """
+        self.perturbation = perturbation
+        self.vectorizer = vectorizer
+        super().__init__(cfg, zarr_dataset)
+
+    def _get_sample_function(self) -> Callable[..., dict]:
+        return partial(
+            generate_agent_sample_vectorized,
+            history_num_frames_ego=self.cfg["model_params"]["history_num_frames_ego"],
+            history_num_frames_agents=self.cfg["model_params"]["history_num_frames_agents"],
+            future_num_frames=self.cfg["model_params"]["future_num_frames"],
+            step_time=self.cfg["model_params"]["step_time"],
+            filter_agents_threshold=self.cfg["raster_params"]["filter_agents_threshold"],
+            perturbation=self.perturbation,
+            vectorizer=self.vectorizer
+        )
+
+    def get_scene_dataset(self, scene_index: int) -> "EgoDatasetVectorized":
+        dataset = super().get_scene_dataset(scene_index).dataset
+        return EgoDatasetVectorized(self.cfg, dataset, self.vectorizer, self.perturbation)
