+++ l5kit/l5kit/sampling/agent_sampling_vectorized.py
@@ -9,7 +9,7 @@ from ..data.filter import filter_agents_by_track_id
 from ..geometry import compute_agent_pose, rotation33_as_yaw
 from ..kinematic import Perturbation
 from ..rasterization import EGO_EXTENT_HEIGHT, EGO_EXTENT_LENGTH, EGO_EXTENT_WIDTH
-from ..sampling import compute_agent_velocity, get_agent_context, get_relative_poses
+from ..sampling.agent_sampling import compute_agent_velocity, get_agent_context, get_relative_poses
 
 
 def generate_agent_sample_vectorized(
